Version 12

+++ frappe/commands/site.py
@@ -1,19 +1,32 @@
-from __future__ import unicode_literals, absolute_import, print_function
+# imports - standard imports
+import atexit
+import compileall
+import hashlib
+import os
+import re
+import shutil
+import sys
+
+# imports - third party imports
 import click
-import hashlib, os, sys, compileall, re
+
+# imports - module imports
 import frappe
 from frappe import _
-from frappe.commands import pass_context, get_site
+from frappe.commands import get_site, pass_context
 from frappe.commands.scheduler import _is_scheduler_enabled
+from frappe.exceptions import SiteNotSpecifiedError
 from frappe.installer import update_site_config
-from frappe.utils import touch_file, get_site_path
-from six import text_type
+from frappe.utils import get_site_path, touch_file
+
 
 @click.command('new-site')
 @click.argument('site')
 @click.option('--db-name', help='Database name')
 @click.option('--db-password', help='Database password')
 @click.option('--db-type', default='mariadb', type=click.Choice(['mariadb', 'postgres']), help='Optional "postgres" or "mariadb". Default is "mariadb"')
+@click.option('--db-host', help='Database Host')
+@click.option('--db-port', type=int, help='Database Port')
 @click.option('--mariadb-root-username', default='root', help='Root username for MariaDB')
 @click.option('--mariadb-root-password', help='Root password for MariaDB')
 @click.option('--no-mariadb-socket', is_flag=True, default=False, help='Set MariaDB host to % and use TCP/IP Socket instead of using the UNIX Socket')
@@ -31,14 +44,16 @@ def new_site(site, mariadb_root_username=None, mariadb_root_password=None, admin
 	_new_site(db_name, site, mariadb_root_username=mariadb_root_username,
 			  mariadb_root_password=mariadb_root_password, admin_password=admin_password,
 			  verbose=verbose, install_apps=install_app, source_sql=source_sql, force=force,
-			  no_mariadb_socket=no_mariadb_socket, db_type=db_type, db_host=db_host, db_password=db_password, db_port=db_port)
+			  no_mariadb_socket=no_mariadb_socket, db_password=db_password, db_type=db_type, db_host=db_host,
+			  db_port=db_port, new_site=True)
 
 	if len(frappe.utils.get_sites()) == 1:
 		use(site)
 
 def _new_site(db_name, site, mariadb_root_username=None, mariadb_root_password=None,
 			  admin_password=None, verbose=False, install_apps=None, source_sql=None, force=False,
-			  no_mariadb_socket=False, reinstall=False, db_type=None, db_host=None, db_password=None, db_port=None):
+			  no_mariadb_socket=False, reinstall=False,  db_password=None, db_type=None, db_host=None,
+			  db_port=None, new_site=False):
 	"""Install a new Frappe site"""
 
 	if not force and os.path.exists(site):
@@ -66,32 +81,23 @@ def _new_site(db_name, site, mariadb_root_username=None, mariadb_root_password=N
 
 	make_site_dirs()
 
-	installing = None
-	try:
-		installing = touch_file(get_site_path('locks', 'installing.lock'))
+	installing = touch_file(get_site_path('locks', 'installing.lock'))
 
-		install_db(root_login=mariadb_root_username, root_password=mariadb_root_password,
-				   db_name=db_name, admin_password=admin_password, verbose=verbose,
-				   source_sql=source_sql, force=force, reinstall=reinstall, db_type=db_type, db_host=db_host, db_port=db_port, db_password=db_password, no_mariadb_socket=no_mariadb_socket)
+	install_db(root_login=mariadb_root_username, root_password=mariadb_root_password, db_name=db_name,
+		admin_password=admin_password, verbose=verbose, source_sql=source_sql, force=force, reinstall=reinstall,
+		db_password=db_password, db_type=db_type, db_host=db_host, db_port=db_port, no_mariadb_socket=no_mariadb_socket)
+	apps_to_install = ['frappe'] + (frappe.conf.get("install_apps") or []) + (list(install_apps) or [])
+	for app in apps_to_install:
+		_install_app(app, verbose=verbose, set_as_patched=not source_sql)
 
-		apps_to_install = ['frappe'] + (frappe.conf.get("install_apps") or []) + (list(install_apps) or [])
-		for app in apps_to_install:
-			_install_app(app, verbose=verbose, set_as_patched=not source_sql)
+	os.remove(installing)
 
-		frappe.utils.scheduler.toggle_scheduler(enable_scheduler)
-		frappe.db.commit()
+	frappe.utils.scheduler.toggle_scheduler(enable_scheduler)
+	frappe.db.commit()
 
-		scheduler_status = "disabled" if frappe.utils.scheduler.is_scheduler_disabled() else "enabled"
-		print("*** Scheduler is", scheduler_status, "***")
+	scheduler_status = "disabled" if frappe.utils.scheduler.is_scheduler_disabled() else "enabled"
+	print("*** Scheduler is", scheduler_status, "***")
 
-	except frappe.exceptions.ImproperDBConfigurationError:
-		_drop_site(site, mariadb_root_username, mariadb_root_password, force=True)
-
-	finally:
-		if installing and os.path.exists(installing):
-			os.remove(installing)
-
-		frappe.destroy()
 
 @click.command('restore')
 @click.argument('sql-file-path')
@@ -102,36 +108,60 @@ def _new_site(db_name, site, mariadb_root_username=None, mariadb_root_password=N
 @click.option('--install-app', multiple=True, help='Install app after installation')
 @click.option('--with-public-files', help='Restores the public files of the site, given path to its tar file')
 @click.option('--with-private-files', help='Restores the private files of the site, given path to its tar file')
+@click.option('--force', is_flag=True, default=False, help='Use a bit of force to get the job done')
 @pass_context
 def restore(context, sql_file_path, mariadb_root_username=None, mariadb_root_password=None, db_name=None, verbose=None, install_app=None, admin_password=None, force=None, with_public_files=None, with_private_files=None):
 	"Restore site database from an sql file"
-	from frappe.installer import extract_sql_gzip, extract_tar_files
-	# Extract the gzip file if user has passed *.sql.gz file instead of *.sql file
+	from frappe.installer import extract_sql_gzip, extract_tar_files, is_downgrade
+	force = context.force or force
 
+	# Extract the gzip file if user has passed *.sql.gz file instead of *.sql file
 	if not os.path.exists(sql_file_path):
-		sql_file_path = '../' + sql_file_path
+		base_path = '..'
+		sql_file_path = os.path.join(base_path, sql_file_path)
 		if not os.path.exists(sql_file_path):
 			print('Invalid path {0}'.format(sql_file_path[3:]))
 			sys.exit(1)
+	elif sql_file_path.startswith(os.sep):
+		base_path = os.sep
+	else:
+		base_path = '.'
 
 	if sql_file_path.endswith('sql.gz'):
-		sql_file_path = extract_sql_gzip(os.path.abspath(sql_file_path))
+		decompressed_file_name = extract_sql_gzip(os.path.abspath(sql_file_path))
+	else:
+		decompressed_file_name = sql_file_path
 
 	site = get_site(context)
 	frappe.init(site=site)
+
+	# dont allow downgrading to older versions of frappe without force
+	if not force and is_downgrade(decompressed_file_name, verbose=True):
+		warn_message = "This is not recommended and may lead to unexpected behaviour. Do you want to continue anyway?"
+		click.confirm(warn_message, abort=True)
+
 	_new_site(frappe.conf.db_name, site, mariadb_root_username=mariadb_root_username,
 		mariadb_root_password=mariadb_root_password, admin_password=admin_password,
-		verbose=context.verbose, install_apps=install_app, source_sql=sql_file_path,
-		force=context.force)
+		verbose=context.verbose, install_apps=install_app, source_sql=decompressed_file_name,
+		force=True, db_type=frappe.conf.db_type)
 
 	# Extract public and/or private files to the restored site, if user has given the path
 	if with_public_files:
+		with_public_files = os.path.join(base_path, with_public_files)
 		public = extract_tar_files(site, with_public_files, 'public')
 		os.remove(public)
 
 	if with_private_files:
+		with_private_files = os.path.join(base_path, with_private_files)
 		private = extract_tar_files(site, with_private_files, 'private')
 		os.remove(private)
+
+	# Removing temporarily created file
+	if decompressed_file_name != sql_file_path:
+		os.remove(decompressed_file_name)
+
+	success_message = "Site {0} has been restored{1}".format(site, " with files" if (with_public_files or with_private_files) else "")
+	click.secho(success_message, fg="green")
 
 @click.command('reinstall')
 @click.option('--admin-password', help='Administrator Password for reinstalled site')
@@ -171,14 +201,31 @@ def _reinstall(site, admin_password=None, mariadb_root_username=None, mariadb_ro
 def install_app(context, apps):
 	"Install a new app to site, supports multiple apps"
 	from frappe.installer import install_app as _install_app
+	exit_code = 0
+
+	if not context.sites:
+		raise SiteNotSpecifiedError
+
 	for site in context.sites:
 		frappe.init(site=site)
 		frappe.connect()
-		try:
-			for app in apps:
+
+		for app in apps:
+			try:
 				_install_app(app, verbose=context.verbose)
-		finally:
-			frappe.destroy()
+			except frappe.IncompatibleApp as err:
+				err_msg = ":\n{}".format(err) if str(err) else ""
+				print("App {} is Incompatible with Site {}{}".format(app, site, err_msg))
+				exit_code = 1
+			except Exception as err:
+				err_msg = ":\n{}".format(err if str(err) else frappe.get_traceback())
+				print("An error occurred while installing {}{}".format(app, err_msg))
+				exit_code = 1
+
+		frappe.destroy()
+
+	sys.exit(exit_code)
+
 
 @click.command('list-apps')
 @pass_context
@@ -208,7 +255,8 @@ def add_system_manager(context, email, first_name, last_name, send_welcome_email
 			frappe.db.commit()
 		finally:
 			frappe.destroy()
-
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('disable-user')
 @click.argument('email')
@@ -239,9 +287,25 @@ def migrate(context, rebuild_website=False, skip_failing=False):
 			migrate(context.verbose, rebuild_website=rebuild_website, skip_failing=skip_failing)
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 	print("Compiling Python Files...")
 	compileall.compile_dir('../apps', quiet=1, rx=re.compile('.*node_modules.*'))
+
+@click.command('migrate-to')
+@click.argument('frappe_provider')
+@pass_context
+def migrate_to(context, frappe_provider):
+	"Migrates site to the specified provider"
+	from frappe.integrations.frappe_providers import migrate_to
+	for site in context.sites:
+		frappe.init(site=site)
+		frappe.connect()
+		migrate_to(site, frappe_provider)
+		frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('run-patch')
 @click.argument('module')
@@ -256,6 +320,8 @@ def run_patch(context, module):
 			frappe.modules.patch_handler.run_single(module, force=context.force)
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('reload-doc')
 @click.argument('module')
@@ -272,6 +338,8 @@ def reload_doc(context, module, doctype, docname):
 			frappe.db.commit()
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('reload-doctype')
 @click.argument('doctype')
@@ -286,6 +354,8 @@ def reload_doctype(context, doctype):
 			frappe.db.commit()
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('add-to-hosts')
 @pass_context
@@ -293,6 +363,8 @@ def add_to_hosts(context):
 	"Add site to hosts"
 	for site in context.sites:
 		frappe.commands.popen('echo 127.0.0.1\t{0} | sudo tee -a /etc/hosts'.format(site))
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('use')
 @click.argument('site')
@@ -305,28 +377,43 @@ def use(site, sites_path='.'):
 		with open(os.path.join(sites_path,  "currentsite.txt"), "w") as sitefile:
 			sitefile.write(site)
 	else:
-		print("{} does not exist".format(site))
+		print("Site {} does not exist".format(site))
 
 @click.command('backup')
 @click.option('--with-files', default=False, is_flag=True, help="Take backup with files")
+@click.option('--verbose', default=False, is_flag=True)
 @pass_context
 def backup(context, with_files=False, backup_path_db=None, backup_path_files=None,
-	backup_path_private_files=None, quiet=False):
+	backup_path_private_files=None, quiet=False, verbose=False):
 	"Backup"
 	from frappe.utils.backups import scheduled_backup
-	verbose = context.verbose
+	verbose = verbose or context.verbose
+	exit_code = 0
 	for site in context.sites:
-		frappe.init(site=site)
-		frappe.connect()
-		odb = scheduled_backup(ignore_files=not with_files, backup_path_db=backup_path_db, backup_path_files=backup_path_files, backup_path_private_files=backup_path_private_files, force=True)
+		try:
+			frappe.init(site=site)
+			frappe.connect()
+			odb = scheduled_backup(ignore_files=not with_files, backup_path_db=backup_path_db, backup_path_files=backup_path_files, backup_path_private_files=backup_path_private_files, force=True, verbose=verbose)
+		except Exception as e:
+			if verbose:
+				print("Backup failed for {0}. Database or site_config.json may be corrupted".format(site))
+			exit_code = 1
+			continue
+
 		if verbose:
 			from frappe.utils import now
-			print("database backup taken -", odb.backup_path_db, "- on", now())
+			summary_title = "Backup Summary at {0}".format(now())
+			print(summary_title + "\n" + "-" * len(summary_title))
+			print("Database backup:", odb.backup_path_db)
 			if with_files:
-				print("files backup taken -", odb.backup_path_files, "- on", now())
-				print("private files backup taken -", odb.backup_path_private_files, "- on", now())
+				print("Public files:   ", odb.backup_path_files)
+				print("Private files:  ", odb.backup_path_private_files)
 
 		frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
+
+	sys.exit(exit_code)
 
 @click.command('remove-from-installed-apps')
 @click.argument('app')
@@ -341,22 +428,28 @@ def remove_from_installed_apps(context, app):
 			remove_from_installed_apps(app)
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('uninstall-app')
 @click.argument('app')
 @click.option('--yes', '-y', help='To bypass confirmation prompt for uninstalling the app', is_flag=True, default=False, multiple=True)
 @click.option('--dry-run', help='List all doctypes that will be deleted', is_flag=True, default=False)
+@click.option('--no-backup', help='Do not backup the site', is_flag=True, default=False)
+@click.option('--force', help='Force remove app from site', is_flag=True, default=False)
 @pass_context
-def uninstall(context, app, dry_run=False, yes=False):
+def uninstall(context, app, dry_run, yes, no_backup, force):
 	"Remove app and linked modules from site"
 	from frappe.installer import remove_app
 	for site in context.sites:
 		try:
 			frappe.init(site=site)
 			frappe.connect()
-			remove_app(app, dry_run, yes)
+			remove_app(app_name=app, dry_run=dry_run, yes=yes, no_backup=no_backup, force=force)
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 
 @click.command('drop-site')
@@ -387,7 +480,7 @@ def _drop_site(site, root_login='root', root_password=None, archived_sites_path=
 		else:
 			click.echo("="*80)
 			click.echo("Error: The operation has stopped because backup of {s}'s database failed.".format(s=site))
-			click.echo("Reason: {reason}{sep}".format(reason=err[1], sep="\n"))
+			click.echo("Reason: {reason}{sep}".format(reason=str(err), sep="\n"))
 			click.echo("Fix the issue and try again.")
 			click.echo(
 				"Hint: Use 'bench drop-site {s} --force' to force the removal of {s}".format(sep="\n", tab="\t", s=site)
@@ -448,6 +541,8 @@ def set_admin_password(context, admin_password, logout_all_sessions=False):
 			admin_password = None
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('set-last-active-for-user')
 @click.option('--user', help="Setup last active date for user")
@@ -493,6 +588,8 @@ def publish_realtime(context, event, message, room, user, doctype, docname, afte
 			frappe.db.commit()
 		finally:
 			frappe.destroy()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 @click.command('browse')
 @click.argument('site', required=False)
@@ -520,6 +617,8 @@ def start_recording(context):
 	for site in context.sites:
 		frappe.init(site=site)
 		frappe.recorder.start()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 
 @click.command('stop-recording')
@@ -528,6 +627,8 @@ def stop_recording(context):
 	for site in context.sites:
 		frappe.init(site=site)
 		frappe.recorder.stop()
+	if not context.sites:
+		raise SiteNotSpecifiedError
 
 
 commands = [
@@ -537,6 +638,7 @@ commands = [
 	install_app,
 	list_apps,
 	migrate,
+	migrate_to,
 	new_site,
 	reinstall,
 	reload_doc,
